added users operations
updateUser deleteUser getallusers
